+++ recipes-core/initscripts/files/imx8mp-irma6r2/set-mount-permissions.sh
@@ -1,7 +1,13 @@
 #!/bin/sh
 
 MOUNTPOINT="/mnt/iris/"
+IDENTITY_DIR="/mnt/iris/identity"
 MOUNT_GROUP="irma_userdata"
 
 # No "chown -R" because we don't want to touch subdirs
-[ "$(stat -c '%g' $MOUNTPOINT)" = 0 ] && chown :"$MOUNT_GROUP" "$MOUNTPOINT" && chmod 0775 "$MOUNTPOINT"
+chown root:"$MOUNT_GROUP" "$MOUNTPOINT" && chmod 0775 "$MOUNTPOINT"
+
+# Restrict access to identity files
+chown root:root -R "$IDENTITY_DIR"
+chmod 0400 -R "$IDENTITY_DIR"
+chmod u+x "$IDENTITY_DIR"